Add loader factory to support both v1 format and old format
