+++ Medidata.Rave.Tsdv.Loader.Sample/Program.cs
@@ -31,7 +31,7 @@ namespace Medidata.Rave.Tsdv.Loader.Sample
 
         private static void UploadTsdvReport(string filePath)
         {
-            var loader = _container.Resolve<ITsdvExcelLoaderFactory>().Create();
+            var loader = _container.Resolve<ITsdvExcelLoaderFactory>().Create(TsdvLoaderSupportedVersion.V1);
 
             Console.WriteLine("Loading from stream");
             using (var fs = new FileStream(filePath, FileMode.Open))
@@ -52,7 +52,7 @@ namespace Medidata.Rave.Tsdv.Loader.Sample
 
         private static void DownloadTsdvReport(string filePath)
         {
-            var loader = _container.Resolve<ITsdvExcelLoaderFactory>().Create();
+            var loader = _container.Resolve<ITsdvExcelLoaderFactory>().Create(TsdvLoaderSupportedVersion.V1);
 
             // Case 1
             // Define a sheet by model type, and add items
