+++ Medidata.Rave.Tsdv.Loader/ITsdvExcelLoaderFactory.cs
@@ -4,6 +4,6 @@ namespace Medidata.Rave.Tsdv.Loader
 {
     public interface ITsdvExcelLoaderFactory
     {
-        IExcelLoader Create();
+        IExcelLoader Create(TsdvLoaderSupportedVersion version);
     }
 }
