+++ Medidata.Rave.Tsdv.Loader/Medidata.Rave.Tsdv.Loader.csproj
@@ -57,9 +57,8 @@
     <Compile Include="SheetDefinitions\v1\TierFolder.cs" />
     <Compile Include="SheetDefinitions\v1\TierForm.cs" />
     <Compile Include="TranslateHeaderDecorator.cs" />
-    <Compile Include="TsdvExcelLoaderFactory.cs" />
-    <Compile Include="TsdvExcelLoaderFactoryBase.cs" />
     <Compile Include="TsdvExcelLoaderOldFormatFactory.cs" />
+    <Compile Include="TsdvExcelLoaderV1Factory.cs" />
     <Compile Include="CoveredExcelBuilder.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Resource.Designer.cs">
@@ -74,6 +73,7 @@
     <Compile Include="SheetDefinitions\OldFormat\TierField.cs" />
     <Compile Include="SheetDefinitions\OldFormat\TierFolder.cs" />
     <Compile Include="SheetDefinitions\OldFormat\TierForm.cs" />
+    <Compile Include="TsdvLoaderSupportedVersion.cs" />
     <Compile Include="Validations\Rules\ValidateBlockPlanSettings.cs" />
     <Compile Include="Validations\Rules\ValidateBlockPlanSettingTiers.cs" />
     <Compile Include="Validations\Rules\LocalizableValidationRuleBase.cs" />
