+++ Medidata.Rave.Tsdv.Loader/Medidata.Rave.Tsdv.Loader.nuspec
@@ -2,14 +2,14 @@
 <package>
   <metadata>
     <id>Medidata.Rave.Tsdv.Loader</id>
-    <version>0.1.9</version>
+    <version>0.1.10</version>
     <title>Medidata.Rave.Tsdv.Loader</title>
     <authors>jshao@mdsol.com,echen@mdsol.com,chuang@mdsol.com</authors>
     <owners>jshao@mdsol.com,echen@mdsol.com,chuang@mdsol.com</owners>
     <projectUrl>https://github.com/junshao/Medidata.Cloud.Tsdv.Loader</projectUrl>
     <requireLicenseAcceptance>false</requireLicenseAcceptance>
     <description>Medidata TSDV Excel report unloading and downloading</description>
-    <releaseNotes>Add sheet definitions for new spreadsheet format.</releaseNotes>
+    <releaseNotes>Add sheet definitions for new spreadsheet format and loader factory to support version 1 format and old format.</releaseNotes>
     <copyright>Copyright 2016</copyright>
     <tags>TSDV parser loader</tags>
     <dependencies>
