+++ Medidata.Rave.Tsdv.Loader/TsdvExcelLoaderV1Factory.cs
@@ -4,11 +4,14 @@ using Medidata.Rave.Tsdv.Loader.SheetDefinitions.v1;
 
 namespace Medidata.Rave.Tsdv.Loader
 {
-    public class TsdvExcelLoaderFactory : TsdvExcelLoaderFactoryBase
+    public class TsdvExcelLoaderV1Factory : TsdvExcelLoaderOldFormatFactory
     {
-        public TsdvExcelLoaderFactory(ILocalization localization) : base(localization) {}
-        protected override IExcelLoader DefineTsdvSheets(IExcelLoader loader)
+        public TsdvExcelLoaderV1Factory(ILocalization localization) : base(localization) {}
+
+        protected override IExcelLoader DefineTsdvSheets(TsdvLoaderSupportedVersion version, IExcelLoader loader)
         {
+            if (version != TsdvLoaderSupportedVersion.V1) return base.DefineTsdvSheets(version, loader);
+
             loader.Sheet<BlockPlanSetting>();
             loader.Sheet<CustomTier>();
             loader.Sheet<TierForm>();
