+++ Medidata.Rave.Tsdv.Loader/TsdvLoaderSupportedVersion.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Medidata.Rave.Tsdv.Loader
+{
+    [Flags]
+    public enum TsdvLoaderSupportedVersion
+    {
+        OldFormat = 0,
+        V1 = 1
+    }
+}
